Commit before rewrite to use JButtonTable

+++ TimeTable/src/degubi/Main.java
@@ -45,12 +45,10 @@ public final class Main extends WindowAdapter implements MouseListener{
 	public static final Image icon = Toolkit.getDefaultToolkit().getImage(Main.class.getClassLoader().getResource("assets/tray.png"));
 	public static final TrayIcon tray = new TrayIcon(icon.getScaledInstance(16, 16, Image.SCALE_SMOOTH));
 	public static final Path dataFilePath = Paths.get("classData.txt");
+	public static final Font bigFont = new Font("TimesRoman", Font.PLAIN, 20);
 	private static final Clip beepBoop = getBeepSound();
 	
 	public static void main(String[] args) throws AWTException, IOException {
-		Font bigFont = new Font("TimesRoman", Font.PLAIN, 20);
-		Main main = new Main();
-		
 		frame.setLayout(null);
 		frame.add(newDayButton("Hétfõ", 60, bigFont));
 		frame.add(newDayButton("Kedd", 230, bigFont));
@@ -60,11 +58,9 @@ public final class Main extends WindowAdapter implements MouseListener{
 		frame.setBounds(0, 0, 1024, 768);
 		frame.setLocationRelativeTo(null);
 		
-		if(!Files.exists(dataFilePath)) {
-			Files.write(Main.dataFilePath, "MONDAY Dimat Elõadás 18:00 20:00 Kongresszusi false".getBytes(StandardCharsets.UTF_8), StandardOpenOption.CREATE);
-		}
+		if(!Files.exists(dataFilePath)) Files.write(Main.dataFilePath, "MONDAY Dimat Elõadás 18:00 20:00 Kongresszusi false".getBytes(StandardCharsets.UTF_8), StandardOpenOption.CREATE);
 		
-		ClassDataButton.reloadData(Files.readAllLines(dataFilePath, StandardCharsets.UTF_8));
+		ClassDataButton.reloadData(Files.readAllLines(dataFilePath));
 		
 		DateTimeFormatter displayTimeFormat = DateTimeFormatter.ofPattern("yyyy MM dd, EEEE HH:mm:ss");
 		JLabel label = new JLabel(LocalDateTime.now().format(displayTimeFormat));
@@ -72,6 +68,7 @@ public final class Main extends WindowAdapter implements MouseListener{
 		label.setBounds(400, 10, 300, 40);
 		label.setFont(bigFont);
 		
+		Main main = new Main();
 		frame.add(newButton("Új Óra Hozzáadása", 840, 650, 150, 60, e -> ButtonEditorGui.showEditorGui(true, new ClassDataButton("MONDAY ÓRANÉV Elõadás 08:00 10:00 Terem false"))));
 		frame.setResizable(false);
 		frame.add(label);
